backend/requirements.txt
Django==4.1.6
django_colorfield==0.8.0
djangorestframework==3.14.0
drf_extra_fields==3.4.1
python-dotenv==1.0.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-filter==22.1
pdfkit==1.0.0